Create Counting the elements 1
Counting the elements
In the video, you saw some of Python's most helpful built-in functions.

One of them can be used to count the number of elements in a variable. You'll see three variables in the following steps:

course_ratings - a dictionary variable containing course names as keys and average ratings as values.
course_completions - a list variable containing the daily number of completions for an individual course.
most_popular_course - a string variable containing the name of a course.
You'll practice applying this function to the three variables!

Instructions 1/3

Use a function to count the number of key-value pairs in course_ratings, storing as a variable called num_courses, then print the variable.