Create Cleaning text data
Cleaning text data
In the video, you saw how to build a custom function that performs a calculation and rounds the results. However, custom functions can be used for any task we expect to repeat! One common example is cleaning text data so that it conforms to specific requirements.

In this exercise, you'll create a function that takes string data and:

Replaces spaces with underscores
Converts to lowercase
Returns the formatted string
Instructions

Define a function called clean_string, which takes an argument called text.
Inside the function, create a variable called no_spaces, which contains the text with spaces replaced by underscores.
Inside the function, create a variable called clean_text, which converts characters in no_spaces to lowercase.
Finish the function by producing clean_text as an output.