Create Debugging code
Debugging code
Developers will inevitably produce errors when writing code—this is completely normal! The ability to read error messages and debug or fix your code is a crucial skill when building or maintaining software.

In this exercise, you have been provided with code that produces errors and will need to identify the cause so that you can resolve these issues.

Instructions
Fix the code provided so that it creates and then prints the sales list without any errors.